reformat BE repo style
