+++ server/db/models/role.js
@@ -1,0 +1,15 @@
+import mongoose from 'mongoose';
+
+const roleSchema = new mongoose.Schema({
+    name: {
+        type: String,
+        required: true
+    },
+    description: {
+        type: String
+    }
+})
+
+const Role = mongoose.model("Role", roleSchema);
+
+export default Role;
