+++ server/server.js
@@ -1,7 +1,7 @@
 import express from "express";
 import 'dotenv/config';
 import connectDB from './db/connectDB.js';
-import UserRouter from './user/router.js';
+import UserRouter from './routers/userRouter.js';
 
 const PORT = 3000;
 const app = express();
