--- server/src/user/router.js
@@ -1,13 +1,0 @@
-import express from "express";
-
-const app = express.Router();
-
-app.get('/');
-
-app.get('/:id');
-
-app.put('/:id');
-
-app.delete('/:id');
-
-export default app;
